actor/pyproject.toml
[tool.poetry]
name = "actor"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "actor"}]

[tool.poetry.scripts]
main = "actor.main:run"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.25.1"
openai = "^0.27.8"
langchain = "^0.0.251"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
